refactor: move base.ts to internal.ts and remove from index
