+++ packages/sfdx-ts-types/src/narrowing/internal.ts
